LayoutTests/fast/mutation/observe-attributes-expected.txt

Test WebKitMutationObserver.observe on attributes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing basic aspects of attribute observation.
...can attribute changes be observed at all
PASS mutations.length is 1
PASS mutations[0].type is "attributes"
PASS mutations[0].attributeName is "foo"
PASS mutations[0].attributeNamespace is null
...observer.disconnect() should prevent further delivery of mutations.
PASS mutations is null
...re-observing after disconnect works with the same observer.
PASS mutations.length is 2
PASS mutations[0].type is "attributes"
PASS mutations[0].attributeName is "foo"
PASS mutations[0].attributeNamespace is null
PASS mutations[1].type is "attributes"
PASS mutations[1].attributeName is "bar"
PASS mutations[1].attributeNamespace is null

Testing that observing without specifying "attributes" does not result in hearing about attribute changes.
PASS mutations is null

Testing that re-observing the same node with the same observer has the effect of resetting the options.
PASS calls is 1
PASS mutations.length is 1
PASS mutations[0].type is "attributes"
PASS mutations[0].attributeName is "foo"
PASS mutations is null

Testing that multiple observers can be registered to a given node and both receive mutations.
PASS mutations.length is 1
PASS mutations[0].type is "attributes"
PASS mutations[0].attributeName is "foo"
PASS mutations2.length is 1
PASS mutations2[0].type is "attributes"
PASS mutations2[0].attributeName is "foo"

Testing that "attributeNamespace" value is delivered properly.
PASS mutations.length is 1
PASS mutations[0].type is "attributes"
PASS mutations[0].attributeName is "foo"
PASS mutations[0].attributeNamespace is "http://www.foo.com/bar"

Testing that modifications to node properties which delegate to attribute storage deliver mutations.
PASS mutations.length is 2
PASS mutations[0].type is "attributes"
PASS mutations[0].attributeName is "src"
PASS mutations[1].type is "attributes"
PASS mutations[1].attributeName is "href"

Testing mutation records are enqueued for attributes before DOMSubtreeModified is dispatched.
PASS mutations.length is 2
PASS mutations[0].type is "attributes"
PASS mutations[0].attributeName is "foo"
PASS mutations[1].type is "attributes"
PASS mutations[1].attributeName is "baz"

PASS successfullyParsed is true

TEST COMPLETE
